split off open loop tests from test_sequencer_test
