+++ BUILD
@@ -367,7 +367,7 @@ cc_test(
     name = "distbench_test_sequencer_test",
     size = "medium",
     srcs = ["distbench_test_sequencer_test.cc"],
-    shard_count = 26,
+    shard_count = 23,
     deps = [
         ":distbench_test_sequencer_tester",
         ":distbench_utils",
@@ -379,6 +379,17 @@ cc_test(
     size = "medium",
     srcs = ["fanout_test.cc"],
     shard_count = 19,
+    deps = [
+        ":distbench_test_sequencer_tester",
+        ":distbench_utils",
+    ],
+)
+
+cc_test(
+    name = "open_loop_test",
+    size = "medium",
+    srcs = ["open_loop_test.cc"],
+    shard_count = 3,
     deps = [
         ":distbench_test_sequencer_tester",
         ":distbench_utils",
